Add tests for FFT wave simulation: 1. create test fixture
- Add fixture for FFT wave simulation tests
- Move private implementation class into local header and make the fixture a friend to unit test data members
- Add test case for angular spatial wave numbers
