+++ ign-marine/src/WaveSimulationFFT2.cc
@@ -16,6 +16,8 @@
 #include "ignition/marine/WaveSimulationFFT2.hh"
 #include "ignition/marine/WaveSpectrum.hh"
 
+#include "WaveSimulationFFT2Impl.hh"
+
 #include <ignition/common.hh>
 
 #include <fftw3.h>
@@ -28,151 +30,7 @@ namespace ignition
 {
 namespace marine
 {
-
-  ///////////////////////////////////////////////////////////////////////////////
-  // WaveSimulationFFT2Impl
-
-  typedef double fftw_data_type;
-  typedef std::complex<fftw_data_type> complex;
-
-  class WaveSimulationFFT2Impl
-  {
-    public: ~WaveSimulationFFT2Impl();
-
-    public: WaveSimulationFFT2Impl(int _N, double _L);
-
-    public: void SetWindVelocity(double _ux, double _uy);
-
-    public: void SetTime(double _time);
-
-    public: void SetLambda(double _lambda);
-
-    public: void ComputeHeights(
-      std::vector<double>& _heights);
-
-    public: void ComputeHeightDerivatives(
-      std::vector<double>& _dhdx,
-      std::vector<double>& _dhdy);
-
-    public: void ComputeDisplacements(
-      std::vector<double>& _sx,
-      std::vector<double>& _sy);
-
-    public: void ComputeDisplacementDerivatives(
-      std::vector<double>& _dsxdx,
-      std::vector<double>& _dsydy,
-      std::vector<double>& _dsxdy);
-
-    private: void ComputeBaseAmplitudes();
-
-    private: void ComputeCurrentAmplitudes(double _time);
-    
-    int mN;
-    int mN2;
-    double mL;
-    double mLambda;
-
-    std::vector<complex> mH;      // FFT0 - height
-    std::vector<complex> mHikx;   // FFT1 - d height / dx
-    std::vector<complex> mHiky;   // FFT1 - d height / dy
-    std::vector<complex> mDx;     // FFT3 - displacement x
-    std::vector<complex> mDy;     // FFT4 - displacement y
-    std::vector<complex> mHkxkx;  // FFT5 - d displacement x / dx
-    std::vector<complex> mHkyky;  // FFT6 - d displacement y / dy
-    std::vector<complex> mHkxky;  // FFT7 - d displacement x / dy = d displacement y / dx
-
-    fftw_complex* mIn0;
-    fftw_complex* mIn1;
-    fftw_complex* mIn2;
-    fftw_complex* mIn3;
-    fftw_complex* mIn4;
-    fftw_complex* mIn5;
-    fftw_complex* mIn6;
-    fftw_complex* mIn7;
-
-    fftw_complex* mOut0;
-    fftw_complex* mOut1;
-    fftw_complex* mOut2;
-    fftw_complex* mOut3;
-    fftw_complex* mOut4;
-    fftw_complex* mOut5;
-    fftw_complex* mOut6;
-    fftw_complex* mOut7;
-
-    fftw_plan mFFTPlan0, mFFTPlan1, mFFTPlan2;
-    fftw_plan mFFTPlan3, mFFTPlan4, mFFTPlan5, mFFTPlan6, mFFTPlan7;
-
-    ////////////////////////////////////////////////////////////
-    /// \note: reworked version
-
-    // parameters
-    double  Lx = this->mL;
-    double  Ly = this->mL;
-    int     Nx = this->mN;
-    int     Ny = this->mN;
-    double  u10 = 5.0;
-    double  phi10 = 0.0;
-    double  s_param = 5.0;
-    double  cap_omega_c = 0.84;
-
-    // derived quantities
-
-    // sample spacing [m]
-    double  delta_x = this->Lx / this->Nx;
-    double  delta_y = this->Ly / this->Ny;
-
-    // fundamental wavelength [m]
-    double  lambda_x_f = this->Lx;
-    double  lambda_y_f = this->Ly;
-
-    // Nyquist wavelength [m]
-    double  lambda_x_Ny = 2.0 * this->delta_x;
-    double  lambda_y_Ny = 2.0 * this->delta_y;
-
-    // fundamental spatial frequency [1/m]
-    double  nu_x_f = 1.0 / this->Lx;
-    double  nu_y_f = 1.0 / this->Ly;
-
-    // Nyquist spatial frequency [1/m]
-    double  nu_x_Ny = 1.0 / (2.0 * this->delta_x);
-    double  nu_y_Ny = 1.0 / (2.0 * this->delta_y);
-
-    // fundamental angular spatial frequency [rad/m]
-    double  kx_f = 2.0 * M_PI / this->Lx;
-    double  ky_f = 2.0 * M_PI / this->Ly;
-
-    // Nyquist angular spatial frequency [rad/m]
-    double  kx_Ny = this->kx_f * this->Nx / 2.0;
-    double  ky_Ny = this->ky_f * this->Ny / 2.0;
-
-    // angular spatial frequencies in fft and math order
-    std::vector<double> kx_fft  = std::vector<double>(this->Nx, 0.0);
-    std::vector<double> ky_fft  = std::vector<double>(this->Ny, 0.0);
-    std::vector<double> kx_math = std::vector<double>(this->Nx, 0.0);
-    std::vector<double> ky_math = std::vector<double>(this->Ny, 0.0);
-
-    // set to 1 to use a symmetric spreading function (=> standing waves)
-    bool use_symmetric_spreading_fn = false;
-
-    // square-root of two-sided discrete elevation variance spectrum
-    std::vector<std::vector<double>> cap_psi_2s_root = std::vector<std::vector<double>>(
-        this->Nx, std::vector<double>(this->Ny, 0.0));
-
-    // iid random normals for real and imaginary parts of the amplitudes
-    std::vector<std::vector<double>> rho = std::vector<std::vector<double>>(
-        this->Nx, std::vector<double>(this->Ny, 0.0));
-    std::vector<std::vector<double>> sigma = std::vector<std::vector<double>>(
-        this->Nx, std::vector<double>(this->Ny, 0.0));
-
-    // angular temporal frequency
-    std::vector<std::vector<double>> omega_k = std::vector<std::vector<double>>(
-        this->Nx, std::vector<double>(this->Ny, 0.0));
-
-    double ECKVOmniDirectionalSpectrum(double k, double u10, double cap_omega_c=0.84);
-    double ECKVSpreadingFunction(double k, double phi, double u10, double cap_omega_c=0.84);
-    double Cos2SSpreadingFunction(double s_param, double phi, double u10, double cap_omega_c=0.84);
-  };
-
+  /////////////////////////////////////////////////
   WaveSimulationFFT2Impl::~WaveSimulationFFT2Impl()
   {
     fftw_destroy_plan(mFFTPlan0);
@@ -203,6 +61,7 @@ namespace marine
     fftw_free(mIn7);
   }
 
+  /////////////////////////////////////////////////
   WaveSimulationFFT2Impl::WaveSimulationFFT2Impl(int _N, double _L) :
     mN(_N),
     mN2(_N * _N),
@@ -250,6 +109,7 @@ namespace marine
     mFFTPlan7 = fftw_plan_dft_2d(mN, mN, mIn7, mOut7, FFTW_BACKWARD, FFTW_ESTIMATE);
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::SetWindVelocity(double _ux, double _uy)
   {
     // Update wind velocity and recompute base amplitudes.
@@ -259,17 +119,20 @@ namespace marine
     ComputeBaseAmplitudes();
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::SetTime(double _time)
   {
     ComputeCurrentAmplitudes(_time);
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::SetLambda(double _lambda)
   {
     mLambda = _lambda;
     ComputeBaseAmplitudes();
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeHeights(
     std::vector<double>& _heights)
   {
@@ -295,6 +158,7 @@ namespace marine
     }
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeHeightDerivatives(
     std::vector<double>& _dhdx,
     std::vector<double>& _dhdy)
@@ -330,6 +194,7 @@ namespace marine
     }
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeDisplacements(
     std::vector<double>& _sx,
     std::vector<double>& _sy)
@@ -365,6 +230,7 @@ namespace marine
     }
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeDisplacementDerivatives(
     std::vector<double>& _dsxdx,
     std::vector<double>& _dsydy,
@@ -410,6 +276,7 @@ namespace marine
     }
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeBaseAmplitudes()
   {
     // 1D axes
@@ -606,6 +473,7 @@ namespace marine
 
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2Impl::ComputeCurrentAmplitudes(double _time)
   {
     // alias
@@ -728,6 +596,7 @@ namespace marine
 
   }
 
+  /////////////////////////////////////////////////
   double WaveSimulationFFT2Impl::ECKVOmniDirectionalSpectrum(
       double k, double u10, double cap_omega_c)
   {
@@ -816,6 +685,7 @@ namespace marine
     return S;
   }
 
+  /////////////////////////////////////////////////
   double WaveSimulationFFT2Impl::ECKVSpreadingFunction(
       double k, double phi, double u10, double cap_omega_c)
   {
@@ -836,6 +706,7 @@ namespace marine
     return cap_phi;
   }
 
+  /////////////////////////////////////////////////
   double WaveSimulationFFT2Impl::Cos2SSpreadingFunction(
       double s, double phi, double u10, double cap_omega_c)
   {
@@ -852,34 +723,40 @@ namespace marine
     return cap_phi;
   }
 
-  ///////////////////////////////////////////////////////////////////////////////
+  /////////////////////////////////////////////////
   // WaveSimulationFFT2
 
+  /////////////////////////////////////////////////
   WaveSimulationFFT2::~WaveSimulationFFT2()
   {
   }
 
+  /////////////////////////////////////////////////
   WaveSimulationFFT2::WaveSimulationFFT2(int _N, double _L) :
     impl(new WaveSimulationFFT2Impl(_N, _L))
   {
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::SetWindVelocity(double _ux, double _uy)
   {
     impl->SetWindVelocity(_ux, _uy);
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::SetTime(double _time)
   {
     impl->SetTime(_time);
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::ComputeHeights(
     std::vector<double>& _h)
   {
     impl->ComputeHeights(_h);    
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::ComputeHeightDerivatives(
     std::vector<double>& _dhdx,
     std::vector<double>& _dhdy)
@@ -887,6 +764,7 @@ namespace marine
     impl->ComputeHeightDerivatives(_dhdx, _dhdy);  
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::ComputeDisplacements(
     std::vector<double>& _sx,
     std::vector<double>& _sy)
@@ -894,6 +772,7 @@ namespace marine
     impl->ComputeDisplacements(_sx, _sy);    
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::ComputeDisplacementDerivatives(
     std::vector<double>& _dsxdx,
     std::vector<double>& _dsydy,
@@ -902,6 +781,7 @@ namespace marine
     impl->ComputeDisplacementDerivatives(_dsxdx, _dsydy, _dsxdy);      
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::ComputeDisplacementsAndDerivatives(
     std::vector<double>& _h,
     std::vector<double>& _sx,
@@ -918,6 +798,7 @@ namespace marine
     impl->ComputeDisplacementDerivatives(_dsxdx, _dsydy, _dsxdy);    
   }
 
+  /////////////////////////////////////////////////
   void WaveSimulationFFT2::SetLambda(double _lambda)
   {
     impl->SetLambda(_lambda);
